requirements.txt
streamlit==1.49.0
streamlit_authenticator-0.2.3
psycopg2==2.9.10
python-dotenv==1.1.1

#pip install -r .\requiements.txt